refactor: 검색어 태그 입력 시 숫자 허용하도록 변경 (#232)
